Fix uninitialized property, aserver instance
Had some

"conditional move or branch depends on uninitialized value",

and digging revealed this. The 'property.py' is modified to
zero-initialize the properties [1].

The newly introduced test cases show that the uninitialized value only
occurred on the aserver instance for some reason.

Without this patch, one of these tests would fail.

The patch is more of a band-aid than anything else, since
zero-initialization or default constructing
sdbusplus::message::object_path is actually not a valid object path [2].

So there needs to be another patch to add some default constructor
there.

While doing this, found that the

- interface.server.hpp.mako
- interface.aserver.hpp.mako

use a different template to define the properties, which is confusing.
Cleaned that up so they declare their properties in the same way.

Tested: Unit tests pass

References:
[1] https://en.cppreference.com/w/cpp/language/zero_initialization
[2] https://dbus.freedesktop.org/doc/dbus-specification.html#message-protocol-marshaling-object-path

Change-Id: I6ec5823f24fe9430546fc119e83e647ae2b3f8c9
Signed-off-by: Alexander Hansen <alexander.hansen@9elements.com>